Move files into src, adjust main.py call and import
