+++ src/blib/encoding/rich_text_encoder.py
