+++ src/blib/exception.py
