+++ src/blib/formatting/abbreviator.py
