+++ src/blib/formatting/richtext.py
