+++ src/blib/resourceid.py
